IOS-4297: Fix merge issues
