+++ Tangem/App/Services/UserTokenListManager/LockedUserTokenListManager.swift
@@ -30,5 +30,5 @@ struct LockedUserTokenListManager: UserTokenListManager {
         result(.success(()))
     }
 
-    func updateServerFromLocalRepository() {}
+    func upload() {}
 }
