+++ Tangem/App/Services/UserTokenListManager/UserTokenListUpdateType.swift
@@ -13,6 +13,4 @@ enum UserTokenListUpdateType {
     case append(_ entries: [StorageEntry])
     case removeBlockchain(_ blockchainNetwork: BlockchainNetwork)
     case removeToken(_ token: Token, in: BlockchainNetwork)
-    case group(_ groupingOption: StorageEntry.V3.Grouping)
-    case sort(_ sortingOption: StorageEntry.V3.Sorting)
 }
