+++ Tangem/App/Services/UserTokensReordering/UserTokensReorderingOptions.swift
@@ -19,3 +19,27 @@ enum UserTokensReorderingOptions {
         case byBalance
     }
 }
+
+// MARK: - Convenience extensions
+
+extension UserTokensReorderingOptions.Grouping {
+    var isGrouped: Bool {
+        switch self {
+        case .none:
+            return false
+        case .byBlockchainNetwork:
+            return true
+        }
+    }
+}
+
+extension UserTokensReorderingOptions.Sorting {
+    var isSorted: Bool {
+        switch self {
+        case .dragAndDrop:
+            return false
+        case .byBalance:
+            return true
+        }
+    }
+}
