+++ Tangem/Modules/OrganizeTokens/ListSection/OrganizeTokensListSectionViewModel.swift
@@ -2,7 +2,7 @@
 //  OrganizeTokensListSectionViewModel.swift
 //  Tangem
 //
-//  Created by Andrey Fedorov on 23.08.2023.
+//  Created by Andrey Fedorov on 06.06.2023.
 //  Copyright © 2023 Tangem AG. All rights reserved.
 //
