+++ Tangem/Modules/OrganizeTokens/OptionsManagement/OrganizeTokensOptionsManager.swift
@@ -11,97 +11,47 @@ import Combine
 import CombineExt
 
 final class OrganizeTokensOptionsManager {
-    private let userTokenListManager: UserTokenListManager
-    private let editedGroupingOption = CurrentValueSubject<OrganizeTokensOptions.Grouping?, Never>(nil)
-    private let editedSortingOption = CurrentValueSubject<OrganizeTokensOptions.Sorting?, Never>(nil)
+    private let userTokensReorderer: UserTokensReordering
+    private let editedGroupingOption = PassthroughSubject<UserTokensReorderingOptions.Grouping, Never>()
+    private let editedSortingOption = PassthroughSubject<UserTokensReorderingOptions.Sorting, Never>()
 
     init(
-        userTokenListManager: UserTokenListManager
+        userTokensReorderer: UserTokensReordering
     ) {
-        self.userTokenListManager = userTokenListManager
-    }
-
-    private static func userTokenListUpdate(
-        currentValue: OrganizeTokensOptions.Grouping,
-        newValue: OrganizeTokensOptions.Grouping?
-    ) -> UserTokenListUpdateType? {
-        guard let newValue = newValue, newValue != currentValue else { return nil }
-
-        return .group(newValue)
-    }
-
-    private static func userTokenListUpdate(
-        currentValue: OrganizeTokensOptions.Sorting,
-        newValue: OrganizeTokensOptions.Sorting?
-    ) -> UserTokenListUpdateType? {
-        guard let newValue = newValue, newValue != currentValue else { return nil }
-
-        return .sort(newValue)
-    }
-
-    private static func userTokenListUpdate(
-        reorderedWalletModelIds: [WalletModel.ID],
-        existingUserTokens: [StorageEntry.V3.Entry],
-        sortingNewValue: OrganizeTokensOptions.Sorting?
-    ) -> UserTokenListUpdateType? {
-        // Re-orderings made as a result of sorting user tokens by balance are ignored
-        guard sortingNewValue != .byBalance else { return nil }
-
-        var existingUserTokensKeyedByWalletModelIds: [WalletModel.ID: StorageEntry.V3.Entry] = [:]
-        var walletModelIdsFromExistingUserTokens: [WalletModel.ID] = []
-
-        for userToken in existingUserTokens {
-            let walletModelId = userToken.walletModelId
-            walletModelIdsFromExistingUserTokens.append(walletModelId)
-            existingUserTokensKeyedByWalletModelIds[walletModelId] = userToken
-        }
-
-        // Checking if any re-ordering has actually taken place
-        guard walletModelIdsFromExistingUserTokens != reorderedWalletModelIds else { return nil }
-
-        let reorderedUserTokens = reorderedWalletModelIds.compactMap { existingUserTokensKeyedByWalletModelIds[$0] }
-
-        assert(
-            reorderedWalletModelIds.count == existingUserTokens.count && reorderedUserTokens.count == existingUserTokens.count,
-            "Method expects a re-ordered, but not modified (by adding/removing some entries) list of wallet model identifiers"
-        )
-
-        return .rewrite(reorderedUserTokens)
+        self.userTokensReorderer = userTokensReorderer
     }
 }
 
 // MARK: - OrganizeTokensOptionsProviding protocol conformance
 
 extension OrganizeTokensOptionsManager: OrganizeTokensOptionsProviding {
-    var groupingOption: AnyPublisher<OrganizeTokensOptions.Grouping, Never> {
+    var groupingOption: AnyPublisher<UserTokensReorderingOptions.Grouping, Never> {
         let editedGroupingOption = editedGroupingOption
-            .compactMap { $0 }
             .eraseToAnyPublisher()
 
-        let groupingOptionFromUserTokenList = userTokenListManager
-            .groupingOptionPublisher
+        let currentGroupingOption = userTokensReorderer
+            .groupingOption
             .prefix(untilOutputFrom: editedGroupingOption)
             .eraseToAnyPublisher()
 
         return [
             editedGroupingOption,
-            groupingOptionFromUserTokenList,
+            currentGroupingOption,
         ].merge()
     }
 
-    var sortingOption: AnyPublisher<OrganizeTokensOptions.Sorting, Never> {
+    var sortingOption: AnyPublisher<UserTokensReorderingOptions.Sorting, Never> {
         let editedSortingOption = editedSortingOption
-            .compactMap { $0 }
             .eraseToAnyPublisher()
 
-        let sortingOptionFromUserTokenList = userTokenListManager
-            .sortingOptionPublisher
+        let currentSortingOption = userTokensReorderer
+            .sortingOption
             .prefix(untilOutputFrom: editedSortingOption)
             .eraseToAnyPublisher()
 
         return [
             editedSortingOption,
-            sortingOptionFromUserTokenList,
+            currentSortingOption,
         ].merge()
     }
 }
@@ -109,56 +59,29 @@ extension OrganizeTokensOptionsManager: OrganizeTokensOptionsProviding {
 // MARK: - OrganizeTokensOptionsEditing protocol conformance
 
 extension OrganizeTokensOptionsManager: OrganizeTokensOptionsEditing {
-    func group(by groupingOption: OrganizeTokensOptions.Grouping) {
+    func group(by groupingOption: UserTokensReorderingOptions.Grouping) {
         editedGroupingOption.send(groupingOption)
     }
 
-    func sort(by sortingOption: OrganizeTokensOptions.Sorting) {
+    func sort(by sortingOption: UserTokensReorderingOptions.Sorting) {
         editedSortingOption.send(sortingOption)
     }
 
     func save(
         reorderedWalletModelIds: [WalletModel.ID]
     ) -> AnyPublisher<Void, Never> {
-        let manager = userTokenListManager
-
-        let groupingOptionPublisher = Just(nil)
-            .append(editedGroupingOption)
-            .withLatestFrom(userTokenListManager.groupingOptionPublisher) { (currentValue: $1, newValue: $0) }
-
-        let sortingOptionPublisher = Just(nil)
-            .append(editedSortingOption)
-            .withLatestFrom(userTokenListManager.sortingOptionPublisher) { (currentValue: $1, newValue: $0) }
-
         return .just
-            .withLatestFrom(groupingOptionPublisher, sortingOptionPublisher, userTokenListManager.userTokensPublisher)
+            .withLatestFrom(groupingOption, sortingOption)
+            .withWeakCaptureOf(self)
             .flatMapLatest { input in
-                let (groupingOption, sortingOption, userTokens) = input
+                let (manager, (grouping, sorting)) = input
 
-                return Deferred {
-                    Future<Void, Never> { promise in
-                        let updates = [
-                            Self.userTokenListUpdate(
-                                currentValue: groupingOption.currentValue,
-                                newValue: groupingOption.newValue
-                            ),
-                            Self.userTokenListUpdate(
-                                currentValue: sortingOption.currentValue,
-                                newValue: sortingOption.newValue
-                            ),
-                            Self.userTokenListUpdate(
-                                reorderedWalletModelIds: reorderedWalletModelIds,
-                                existingUserTokens: userTokens,
-                                sortingNewValue: sortingOption.newValue
-                            ),
-                        ].compactMap { $0 }
-
-                        DispatchQueue.main.async {
-                            manager.update(updates, shouldUpload: true)
-                            promise(.success(()))
-                        }
-                    }
-                }
+                // TODO: Andrey Fedorov - Add actual implementation (IOS-4297)
+                return manager.userTokensReorderer.reorder([
+                    .setGroupingOption(option: grouping),
+                    .setSortingOption(option: sorting),
+                    .reorder(reorderedWalletModelIds: reorderedWalletModelIds),
+                ])
             }
             .eraseToAnyPublisher()
     }
